BRN-665 ~ Migrated some composer dependencies.
The auto-loading within the project has also been modified to load
specifically the project code and not the common. The common project
has been required as a dependency as an internal link. Essentially the
common composer.json file is being used to required dependencies and
auto-load the commmon code base.