BRN-665 ~ Removed `CSRF` protection.
Long story short the new projects have `csrf_protection` disabled in the
framework by default and it turns out that this being disabled will not
register the form extension that adds this option to the form. Simply
toggling this off removes a bunch of form listeners and actually speeds
form handling as its no longer looking at the session.

Good times 💪