BRN-765 ~ Job has a client collection now.
Other changes relate to naming of the client entities for more verbose
and consistent naming. For example `ShopInterface` is now renamed to
`ClientShopInterface`.

Job and client are joined through a table `JobClient` that allows us to
keep a history of clients that worked on a job. This also means all the
code using the getters/setters is broken. Oh well 💅