BRN-855, BRN-856 ~ Created a new worker called cron that will handle the running of simplistic tasks in an async manor. Task executors are registered in the container and will be come available under the configured name.
Email sending has been refactored to use another table so emails can
be audited and handle via a cron task worker. The emails are also now
sent using a campaign monitor adapter.

A welcome email is sent when a user is created. At the moment a hashed
password is given to the email and set against the user, this will need
changing to satisfy the security needs.

An email is triggered when the user uses a forgotten account API. This
endpoint looks like it can be abused at the moment and needs locking
down somehow, maybe a way of throttling the recovery tokens created.